test(ui): extract shared BrowserHarness base
Move the generic DOM/wait/interaction plumbing (container, q, byRoleName,
buttonByText, inputByName, containsText, waitFor(+Text/+Button/+Transition),
clickElement/Button/MenuItem) and opt-in MSW request tracking into an abstract
BrowserHarness<TFixture>. ProvidersPageHarness now extends it and keeps only its
domain vocabulary. Behaviour unchanged.
